renaming bernstein
